[fix] Move the algorithm for initialize the board and move cells
